Deleted old forecast wrapper
Removed old forecast wrapper and demo, renamed forecastwrapperV2 to
forecastwrapper
